refactor(cli): extract Meta Console REPL handler
